after long press, the goalitem will be deleted

+++ Components/GoalItem.js
@@ -1,23 +1,43 @@
-import { Pressable, StyleSheet, Text, View } from "react-native";
+import { Pressable, StyleSheet, Text, View, Alert } from "react-native";
 import React from "react";
 import { useNavigation } from "@react-navigation/native";
 import PressableButton from "./PressableButton";
-import { MaterialIcons } from '@expo/vector-icons';
+import { MaterialIcons } from '@expo/vector-icons'; 
 
 export default function GoalItem({ goalObj, deleteHandler }) {
   const navigation = useNavigation();
 
   function handleDelete() {
-    console.log("deleted");
-    deleteHandler(goalObj.id);
+    Alert.alert(
+      "Delete Goal",
+      `Are you sure you want to delete this item?`,
+      [
+        {
+          text: "No",
+          style: "cancel",
+        },
+        {
+          text: "Yes",
+          onPress: () => {
+            console.log("Goal deleted");
+            deleteHandler(goalObj.id);
+          },
+          style: "destructive",
+        },
+      ]
+    );
   }
+
   function handlePress() {
+    // Navigate to the Details screen with goal data
     navigation.navigate("Details", { goalData: goalObj });
   }
+
   return (
     <View style={styles.textContainer}>
       <Pressable
         onPress={handlePress}
+        onLongPress={handleDelete}
         style={({ pressed }) => {
           return [styles.horizontalContainer, pressed && styles.pressedStyle];
         }}
@@ -25,7 +45,7 @@ export default function GoalItem({ goalObj, deleteHandler }) {
       >
         <Text style={styles.text}>{goalObj.text}</Text>
 
-        {/* Icon for delete button */}
+        {/* Delete button icon */}
         <PressableButton
           componentStyle={styles.deleteButton}
           pressedHandler={handleDelete}
